+++ src/resources/tenants/index.ts
@@ -1,0 +1,9 @@
+// File generated from our OpenAPI spec by Stainless. See CONTRIBUTING.md for details.
+
+export {
+  Settings,
+  type SettingRetrieveResponse,
+  type SettingUpdateResponse,
+  type SettingUpdateParams,
+} from './settings';
+export { Tenants, type TenantDeleteDataResponse } from './tenants';